HZN-988: Add location-awaraness to path outages, allowing remote critical paths to be tested.
Also refactored parts of the critical path code to use object instead of strings/arrays.
