+++ opennms-dao/src/test/java/org/opennms/netmgt/dao/PathOutageDaoIT.java
@@ -41,10 +41,6 @@ import org.junit.runner.RunWith;
 import org.opennms.core.spring.BeanUtils;
 import org.opennms.core.test.OpenNMSJUnit4ClassRunner;
 import org.opennms.core.test.db.annotations.JUnitTemporaryDatabase;
-import org.opennms.netmgt.dao.api.DistPollerDao;
-import org.opennms.netmgt.dao.api.EventDao;
-import org.opennms.netmgt.dao.api.IpInterfaceDao;
-import org.opennms.netmgt.dao.api.MonitoredServiceDao;
 import org.opennms.netmgt.dao.api.MonitoringLocationDao;
 import org.opennms.netmgt.dao.api.NodeDao;
 import org.opennms.netmgt.dao.api.PathOutageDao;
@@ -79,8 +75,6 @@ import org.springframework.transaction.support.TransactionTemplate;
 @JUnitConfigurationEnvironment
 @JUnitTemporaryDatabase
 public class PathOutageDaoIT implements InitializingBean {
-    @Autowired
-    private DistPollerDao m_distPollerDao;
 
     @Autowired
     private MonitoringLocationDao m_locationDao;
@@ -89,19 +83,10 @@ public class PathOutageDaoIT implements InitializingBean {
     private NodeDao m_nodeDao;
 
     @Autowired
-    private IpInterfaceDao m_ipInterfaceDao;
-
-    @Autowired
-    private MonitoredServiceDao m_monitoredServiceDao;
-
-    @Autowired
     private PathOutageDao m_pathOutageDao;
 
     @Autowired
     private ServiceTypeDao m_serviceTypeDao;
-
-    @Autowired
-    private EventDao m_eventDao;
 
     @Autowired
     TransactionTemplate m_transTemplate;
