+++ opennms-services/src/main/resources/META-INF/opennms/applicationContext-pollerd.xml
@@ -27,7 +27,7 @@
       <property name="name" value="OpenNMS.Poller.DefaultPollContext" />
       <property name="pollerConfig" ref="pollerConfig" />
       <property name="queryManager" ref="pollerQueryManager" />
-      <property name="pingerFactory" ref="pingerFactory" />
+      <property name="locationAwarePingClient" ref="locationAwarePingClient" />
     </bean> 
     
     <bean name="pollableNetwork" class="org.opennms.netmgt.poller.pollables.PollableNetwork">
