+++ opennms-services/src/test/java/org/opennms/netmgt/poller/PollerIT.java
@@ -33,6 +33,7 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
+import static org.mockito.Mockito.mock;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -67,6 +68,7 @@ import org.opennms.netmgt.dao.mock.MockEventIpcManager;
 import org.opennms.netmgt.dao.mock.MockEventIpcManager.SendNowHook;
 import org.opennms.netmgt.eventd.AbstractEventUtil;
 import org.opennms.netmgt.events.api.EventConstants;
+import org.opennms.netmgt.icmp.proxy.LocationAwarePingClient;
 import org.opennms.netmgt.mock.MockElement;
 import org.opennms.netmgt.mock.MockEventUtil;
 import org.opennms.netmgt.mock.MockInterface;
@@ -148,6 +150,8 @@ public class PollerIT implements TemporaryDatabaseAware<MockDatabase> {
     @Autowired
     private LocationAwarePollerClient m_locationAwarePollerClient;
 
+    private LocationAwarePingClient m_locationAwarePingClient;
+
     //
     // SetUp and TearDown
     //
@@ -217,12 +221,15 @@ public class PollerIT implements TemporaryDatabaseAware<MockDatabase> {
         m_eventMgr.setSynchronous(false);
         m_eventMgr.setNumSchedulerThreads(2);
 
+        m_locationAwarePingClient = mock(LocationAwarePingClient.class);
+
         DefaultPollContext pollContext = new DefaultPollContext();
         pollContext.setEventManager(m_eventMgr);
         pollContext.setLocalHostName("localhost");
         pollContext.setName("Test.DefaultPollContext");
         pollContext.setPollerConfig(m_pollerConfig);
         pollContext.setQueryManager(m_queryManager);
+        pollContext.setLocationAwarePingClient(m_locationAwarePingClient);
 
         PollableNetwork network = new PollableNetwork(pollContext);
